成绩

+++ student_server/src/main/java/com/auggie/student_server/entity/StudentCourseTeacher.java
@@ -21,7 +21,7 @@ public class StudentCourseTeacher {
     private String courseId;       // 课程ID，对应数据库中的 course_id
     private String staffId;        // 教师ID，对应数据库中的 staff_id
     private String semester;       // 学期，对应数据库中的 semester
-    private Integer normalScore;   // 平时成绩，对应数据库中的 normal_score
-    private Integer testScore;     // 考试成绩，对应数据库中的 test_score
-    private Integer totalScore;   // 总成绩，对应数据库中的 total_score
+    private Float normalScore;   // 平时成绩，对应数据库中的 normal_score
+    private Float testScore;     // 考试成绩，对应数据库中的 test_score
+    private Float totalScore;   // 总成绩，对应数据库中的 total_score
 }
